move ssh_keyscan from lock.py to misc.py
Because it may be used by provision.py
